fix microsoft Code Standards
