+++ microsoft/src/Pizza.rs
@@ -1,8 +1,11 @@
+
+#[allow(dead_code)]
 pub struct Pizza{
     pub topping: String,
     pub inches: u8,
 }
 
+#[allow(dead_code)]
 impl Pizza{
     pub fn peperoni(inches: u8) -> Self{
         Pizza::bake("peperoni", inches)
