+++ microsoft/src/Result.rs
@@ -1,7 +1,7 @@
 #[derive(Debug)]
 struct OtherError;
 
-
+#[allow(dead_code)]
 fn sum(numa: f64, numb: f64) -> Result<f64, OtherError> {
     if numb == 0.0 {
         Err(OtherError)
