+++ microsoft/src/util_test.rs
@@ -1,3 +1,4 @@
+#[allow(dead_code)]
 fn add(a: i32, b: i32) -> i32 {
     a + b
 }
@@ -8,6 +9,7 @@ fn main(){
     assert_eq!(add(7,9), 16);
 }
 
+#[allow(dead_code)]
 pub fn is_eve(num : i32) -> bool{
     num % 2 == 0
 }
@@ -22,6 +24,7 @@ mod test{
     }
 
     // error
+    #[allow(dead_code)]
     fn is_false_fun(){
         assert!(is_eve(5));
     }
